+++ src/multi_amr/data/postprocessing_str.py
@@ -1,11 +1,10 @@
 import re
 from typing import List
 
-
 def _is_url(text: str) -> bool:
-    # Taken from django https://github.com/django/django/blob/stable/1.3.x/django/core/validators.py#L45
+    # Modified from django https://github.com/django/django/blob/stable/1.3.x/django/core/validators.py#L45
     match = re.match(
-        r'^(?:http|ftp)s?://' # http:// or https://
+        r'^(?:(?:http|ftp)s?://)?' # http:// or https:// (bv: made this optional)
         r'(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}\.?|[A-Z0-9-]{2,}\.?)|' #domain...
         r'localhost|' #localhost...
         r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})' # ...or ip
@@ -21,6 +20,11 @@ def _is_email(text: str) -> bool:
 
     return match is not None
 
+def _is_filename(text: str) -> bool:
+    # Taken from https://stackoverflow.com/a/201378/1150683
+    match = re.match(r"^.*?\.[a-z0-9]+$", text, re.IGNORECASE)
+
+    return match is not None
 
 def clean_up_amr_tokenization(out_string: str) -> str:
     """Clean up a list of simple English tokenization artifacts like spaces before punctuations and abbreviated forms.
@@ -55,7 +59,7 @@ def postprocess_str_after_linearization(linearized: str, verbose: bool = False) 
     # But do not replace _ inside of URLs
     def replace_literal(match):
         content = match.group(1)
-        if not _is_url(content) and not _is_email(content):
+        if not (_is_url(content) or _is_email(content) or _is_filename(content)):
             content = content.replace("_", " ")
         return f"<lit> {content} </lit>"
     linearized = re.sub(r'"(.*?)"', replace_literal, linearized)
@@ -81,7 +85,9 @@ def postprocess_str_after_delinearization(delinearized: str) -> str:
 
     def reverse_literal(match):
         rel = match.group(1).strip()
-        content = match.group(2).strip().replace(" ", "_") if rel.startswith(("wiki", "op")) else match.group(2).strip()
+        content = match.group(2).strip()
+        if rel.startswith(("wiki", "op")):
+            content = content.replace(" ", "_")
 
         return f':{rel} "{content}"'
